% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ladders.R
\name{affix}
\alias{affix}
\title{Affix the session .Rscalade file to use via options()}
\usage{
affix(.scalade)
}
\arguments{
\item{.scalade}{Path to existing .Rscalade file.}
}
\value{
None; called for the side effect of setting the location of an existing .Rscalade
file via \code{options(scalade.rscalade = "/path/to/.Rscalade")}.
}
\description{
A little helper that does the hard part of writing out
\code{options(scalade.rscalade = "/path/to/.Rscalade")} so that all
functions know where to look for the .Rscalade. Overwrites any previously
set .Rscalade in options.
}
