% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_rscalade.R
\name{check}
\alias{check}
\alias{.check_rscalade}
\title{Mega helper to suss out the rscalade location}
\usage{
.check_rscalade(.scalade = NULL)
}
\arguments{
\item{.scalade}{Path to an existing .Rscalade definition. If not provided, walks through inference
steps.}
}
\description{
Checks for a ladder definition. See details.
}
\details{
This function looks for a .Rscalade file as follows:
\itemize{
\item first checks the \code{.scalade} path if provided
\itemize{
\item note this does NOT set \code{options(scalade.rscalade)} to allow for scalade switching;
use \code{scalade::affix()} for that
}
\item if .scalade is not provided, then the following occur in order to infer the location:
\itemize{
\item checks the global options for a preset rscalade path, \code{getOption("scalade.rscalade")}
\item checks for a project .Rscalade (if found, this sets \code{options(scalade.rscalade = .local_scalade)})
\item checks for a home .Rscalade (if found, this sets \code{options(scalade.rscalade = .home_scalade)})
}
}
}
