% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_rscalade.R
\name{.not_alerted}
\alias{.not_alerted}
\title{Helper to check for .Rscalade in project, user home, or specified}
\usage{
.not_alerted()
}
\description{
Alerts regarding if .Rscalade location was inferred
}
\details{
Only allows for the alert to happen once in a session to prevent yelling.
And furthermore only in interactive mode.
}
