% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ladders.R
\name{edit_ladders}
\alias{edit_ladders}
\alias{build_ladder}
\alias{burn_ladder}
\title{Edit the ladder definitions}
\usage{
edit_ladders(.scalade = NULL)

build_ladder(id, repo, packages, .scalade = NULL)

burn_ladder(id.spec = NULL, overwrite = FALSE, .scalade = NULL)
}
\arguments{
\item{id.spec}{Specified ladder ID to work with.}
}
\value{
None; called for the side effect of editing the ~/.Rscalade file.
}
\description{
Add, remove, and edit existing ladder definitions.
}
\details{
Editing of ladders ---------------------------------------------------------

Repos
Note that for packages that are stored in remote (Github) repositories,
the package name must be in a format compatible with \code{remotes::install_remote},
e.g. \verb{<gh-username>/<repo>}.
}
