% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ladders.R
\name{install}
\alias{install}
\alias{transport}
\title{Install & upgrade packages from ladders}
\usage{
transport(id.spec = NULL, .scalade = NULL, ...)
}
\arguments{
\item{id.spec}{Specified ladder ID to inspect. If not specified shows
all available ladders with their associated packages.}

\item{.scalade}{Location of the ladder definition file. Default is NULL.
If NULL (e.g. not set), it first searches for a .Rscalade file in the
project root (as defined by there \code{here} package), and failing that, will
search the user's home directory ("~/.Rscalade").}

\item{...}{Additional arguments to \verb{remotes::install_*}.}
}
\description{
Keep your ladder up to date by moving from a source repo the latest
version of a package to your computer (hence, transport).
}
\details{
Repos
Note that at the moment this does not take into account if a package
is in two or more repos and will install first from Bioconductor, then CRAN,
then Remote (Github) repositories.
}
