% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ladders.R
\name{install_ladders}
\alias{install_ladders}
\alias{transport_ladder}
\title{Install & upgrade packages from ladders}
\usage{
transport_ladder(id.spec = NULL, .scalade = NULL, ...)
}
\arguments{
\item{id.spec}{Ladders for which to install/upgrade.}

\item{...}{Additional arguments to \verb{remotes::install_*}.}
}
\description{
Keep your ladder up to date by moving from a source repo the latest
version of a package to your computer (hence, transport).
}
\details{
Upgrades ------------------------------------------------------------------

Repos
Note that at the moment this does not take into account if a package
is in two or more repos and will install first from Bioconductor, then CRAN,
then Remote (Github) repositories.
}
