% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ladders.R
\name{load_ladders}
\alias{load_ladders}
\alias{climb_ladder}
\title{Attaches packages specified by ladders}
\usage{
climb_ladder(id.spec, report_conflicts = FALSE, .scalade = NULL)
}
\arguments{
\item{id.spec}{A string or character vector with the ladder IDs
from which to get the packages to load based on the definitions in
~/.Rscalade.}

\item{report_conflicts}{Whether to show the function conflicts that
arise from the loaded packages. Currently set by default to FALSE
as it can be quite noisy.}
}
\description{
Loads up the family of packages associated with the designated
ladders.
}
\details{
Attach ---------------------------------------------------------------------
}
