% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalade.R
\docType{package}
\name{scalade-package}
\alias{scalade}
\alias{scalade-package}
\title{scalade: Ladders for SingleCellExperiments.}
\description{
Helpers to work with SingleCellExperiments. A series of ladders.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/robertamezquita/scalade}
  \item Report bugs at \url{https://github.com/robertamezquita/scalade/issues}
}

}
\author{
\strong{Maintainer}: Robert Amezquita \email{robert.amezquita@me.com} (\href{https://orcid.org/0000-0001-6868-7193}{ORCID})

}
\keyword{internal}
