% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conflicts.R
\name{scalade_conflicts}
\alias{scalade_conflicts}
\title{Conflicts between the scalade and other packages}
\usage{
scalade_conflicts(core, include_internal_conflicts = FALSE)
}
\description{
This function lists all the conflicts between packages in the scalade
and other packages that you have loaded.
}
