% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo.R
\name{scalade_logo_text}
\alias{scalade_logo_text}
\title{The scalade logo, using ASCII or Unicode characters}
\usage{
scalade_logo_text(unicode = l10n_info()$`UTF-8`)
}
\arguments{
\item{unicode}{Whether to use Unicode symbols. Default is \code{TRUE}
on UTF-8 platforms.}
}
\description{
Use \code{\link[crayon:strip_style]{crayon::strip_style()}} to get rid of the colors.
}
\examples{
scalade_logo()
}
