% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scalade_packages}
\alias{scalade_packages}
\title{List all packages in the scalade}
\usage{
scalade_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include scalade in the list?}
}
\description{
List all packages in the scalade
}
\examples{
scalade_packages()
}
