% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scalade_prebuilt}
\alias{scalade_prebuilt}
\title{A series of predefined ladders}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 52 rows and 3 columns.
}
\usage{
scalade_prebuilt
}
\description{
Just some ladders that may be handy.
}
\keyword{datasets}
