% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ladders.R
\name{setup}
\alias{setup}
\title{Creating a fresh .Rscalade definition file}
\usage{
setup(mode = "home", overwrite = FALSE, .scalade = NULL)
}
\arguments{
\item{mode}{One of three options. "home" will create a .Rscalade in your user home
directory. "local" will create a project specific .Rscalade (where your project
location is defined by \code{here::here()}). "manual" will use the provided .scalade
argument.}

\item{overwrite}{If there is an existing .Rscalade file, whether to overwrite it
with the predefined ladder definitions from this package.}

\item{.scalade}{Manually provided desired path for the fresh .Rscalade.}
}
\value{
None; called for side-effect of writing ~/.Rscalade
}
\description{
This function should be the first one run out of the box, as it creates a
baseline definition file that contains predefined ladders.
}
\details{
This package relies on having an available ladder definition file
(.Rscalade). Thus, to use the package, a starting version must be created
via \code{scalade::fresh()}.
}
