% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ladders.R
\name{setup_ladders}
\alias{setup_ladders}
\title{Getting started - creating a ~/.Rscalade definition file}
\usage{
setup_ladders(overwrite = FALSE, .scalade = NULL)
}
\arguments{
\item{overwrite}{Whether to overwrite an existing ~/.Rscalade file with
the original base predefined set of ladders from this package.}
}
\value{
None; called for side-effect of writing ~/.Rscalade
}
\description{
This function should be the first one run out of the box, as it creates a
baseline definition file that contains predefined ladders.
}
\details{
Definitions
This package relies on a centrally cached definitions file that is created
upon first setting up the package with \code{setup_ladders()} .
}
