% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ladders.R
\name{view_ladders}
\alias{view_ladders}
\alias{inspect_ladders}
\title{View defined ladders in a pretty-printed manner}
\usage{
inspect_ladders(id.spec = NULL, .scalade = NULL)
}
\arguments{
\item{id.spec}{Specified ladder ID to inspect. If not specified shows
all available ladders with their associated packages.}
}
\value{
Invisibly returns the tibble of ladder definitions. In the
printed output, a checkmark notes that the package is already installed
on your host computer, while a cross signifies the package is not
currently available.
}
\description{
This will pretty-print ladders (all by default, or specific ones)
that have already been defined in ~/.Rscalade.
}
\details{
Pretty-printing ---------------------------------------------------------
}
